test(exoflex): convert tests for TimePicker and TimePickerInput
